Add images, boilerplate, and basic layout
Added the overall design image and the leafy background image to the
project. Adjusted the README file to credit the person who took the
photo used in the project. Added boilerplate to html, linked css,
and added main with divs. Added scss file and added meyer reset
boilerplate for styling. Added additional styling to main section in
order to set it up for future completion