Add helper files and complete image section
Added fonts folder with necessary odin fonts. Added odin logo png.
Adjusted html to allocate space for the odin logo area and credit to
the photographer. Adjusted styling of links and added necessary links to
use the norsebold font. Added various stylings to the divs in the photo
section of the page in order to make it look like the prompt